Refactor api to move plink support to its own .h/.cc files and remove reference_genome parameter from GenomicsDB constructor
