+++ tools/src/gt_mpi_gather.cc
@@ -798,11 +798,8 @@ int main(int argc, char *argv[]) {
         print_calls(qp, query_config, command_idx, query_config.get_vid_mapper());
         break;
       case COMMAND_PLINK:
-        GenomicsDB gdb(query_config.get_workspace(my_world_mpi_rank),
-                       query_config.get_callset_mapping_file(),
-                       query_config.get_vid_mapping_file(),
-                       query_config.get_reference_genome());
-        gdb.generate_plink(array_name, &query_config, plink_formats, bgen_compression, one_pass, bgen_verbose, progress_interval, plink_prefix, fam_list);
+        GenomicsDB gdb(json_config_file, GenomicsDB::JSON_FILE, loader_json_config_file, my_world_mpi_rank);
+        gdb.generate_plink(plink_formats, bgen_compression, one_pass, bgen_verbose, progress_interval, plink_prefix, fam_list);
         break;
     }
 #ifdef USE_GPERFTOOLS_HEAP
